frontend/requirements.txt
Flask==2.1.0
gunicorn
jaeger-client
prometheus-client
opentelemetry-api
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation
opentelemetry-sdk
prometheus-flask-exporter
pylint
requests
Flask-Opentracing